+++ kernel/framework/io/db/dbms/Doctrine/DBAL/Types/FloatType.php
@@ -5,8 +5,9 @@
  * @subpackage  DBAL\Types
  * @license     https://www.gnu.org/licenses/lgpl-2.1.fr.html LGPL 2.1
  * @link        https://www.doctrine-project.org
- * @version     PHPBoost 6.0 - last update: 2013 01 01
+ * @version     PHPBoost 6.1 - last update: 2025 11 27
  * @since       PHPBoost 4.0 - 2013 01 01
+ * @contributor Sebastien LARTIGUE <babsolune@phpboost.com>
 */
 
 class FloatType extends Type
@@ -23,7 +24,7 @@ class FloatType extends Type
 
     public function convertToPHPValue($value, AbstractPlatform $platform)
     {
-        return (double) $value;
+        return (float) $value;
     }
 }
